Add files via upload
This hackPass.js uses single thread that compares the hashed passwords from jason.2K.hashes.txt (peers hashed passwords) with passwords generated from a password generator to see if there are any matches. if there is a match, the output gets sent to jason.2K.hashes.answers.txt with each line containing the hashed password with the respective password in plain text on the same line. Matches passwords will have the password in plain text or has an space with an empty string. The ones that didn't match doesn't have a space after the hash.